Refactor serialization: split by granularity.
Rather than splitting by algorithm used to serialize a model, it is better to split by granularity, as this can reveal duplicated code that can be simplified. In this change, I only move code around, with minimal changes, to make it easier to review. Next changes will start removing duplicated code, extracting common utilities around, simplifying things by a lot.
