+++ model_signing/serialization/serialize_by_file_shard.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Model serializers that build a single hash out of a DFS traversal."""
+"""Model serializers that operated at file shard level of granularity."""
 
 import base64
 import concurrent.futures
@@ -24,130 +24,43 @@ from model_signing.hashing import file
 from model_signing.hashing import hashing
 from model_signing.manifest import manifest
 from model_signing.serialization import serialization
+from model_signing.serialization import serialize_by_file
 
 
 _ShardSignTask: TypeAlias = tuple[pathlib.PurePath, str, int, int]
-
-
-def check_file_or_directory(path: pathlib.Path) -> None:
-    """Checks that the given path is either a file or a directory.
-
-    There is no support for sockets, pipes, or any other operating system
-    concept abstracted as a file.
-
-    Furthermore, this would raise if the path is a broken symlink, if it doesn't
-    exists or if there are permission errors.
-
-    Args:
-        path: The path to check.
-
-    Raises:
-        ValueError: The path is neither a file or a directory.
-    """
-    if not (path.is_file() or path.is_dir()):
-        raise ValueError(
-            f"Cannot use '{path}' as file or directory. It could be a"
-            " special file, it could be missing, or there might be a"
-            " permission issue."
-        )
 
 
 def _build_header(
     *,
     entry_name: str,
     entry_type: str,
-    start: int | None = None,
-    end: int | None = None,
+    start: int,
+    end: int,
 ) -> bytes:
     """Builds a header to encode a path with given name and type.
 
     Args:
         entry_name: The name of the entry to build the header for.
         entry_type: The type of the entry (file or directory).
-        start: Optional offset for the start of the path shard.
-        end: Optional offset for the end of the path shard.
+        start: Offset for the start of the path shard.
+        end: Offset for the end of the path shard.
 
     Returns:
         A sequence of bytes that encodes all arguments as a sequence of UTF-8
         bytes. Each argument is separated by dots and the last byte is also a
         dot (so the file digest can be appended unambiguously).
     """
+    # Note: This will get replaced in subsequent change, right now we're just
+    # moving existing code around.
     encoded_type = entry_type.encode("utf-8")
     # Prevent confusion if name has a "." inside by encoding to base64.
     encoded_name = base64.b64encode(entry_name.encode("utf-8"))
-
-    if start is not None and end is not None:
-        # Note: make sure to end with a ".".
-        encoded_range = f"{start}-{end}.".encode("utf-8")
-    else:
-        # Note: no "." at end here, it will be added by `join` on return.
-        encoded_range = b""
-
-    return b".".join([encoded_type, encoded_name, encoded_range])
+    encoded_range = f"{start}-{end}".encode("utf-8")
+    # Note: empty string at the end, to terminate header with a "."
+    return b".".join([encoded_type, encoded_name, encoded_range, b""])
 
 
-class DFSSerializer(serialization.Serializer):
-    """Serializer for a model that performs a traversal of the model directory.
-
-    This serializer produces a single hash for the entire model. If the model is
-    a file, the hash is the digest of the file. If the model is a directory, we
-    perform a depth-first traversal of the directory, hash each individual files
-    and aggregate the hashes together.
-    """
-
-    def __init__(
-        self,
-        file_hasher: file.SimpleFileHasher,
-        merge_hasher_factory: Callable[[], hashing.StreamingHashEngine],
-    ):
-        """Initializes an instance to serialize a model with this serializer.
-
-        Args:
-            hasher: The hash engine used to hash the individual files.
-            merge_hasher_factory: A callable that returns a
-              `hashing.StreamingHashEngine` instance used to merge individual
-              file digests to compute an aggregate digest.
-        """
-        self._file_hasher = file_hasher
-        self._merge_hasher_factory = merge_hasher_factory
-
-    @override
-    def serialize(self, model_path: pathlib.Path) -> manifest.DigestManifest:
-        # TODO: github.com/sigstore/model-transparency/issues/196 - Add checks
-        # to exclude symlinks if desired.
-        check_file_or_directory(model_path)
-
-        if model_path.is_file():
-            self._file_hasher.set_file(model_path)
-            return manifest.DigestManifest(self._file_hasher.compute())
-
-        return manifest.DigestManifest(self._dfs(model_path))
-
-    def _dfs(self, directory: pathlib.Path) -> hashing.Digest:
-        # TODO: github.com/sigstore/model-transparency/issues/196 - Add support
-        # for excluded files.
-        children = sorted([x for x in directory.iterdir()])
-
-        hasher = self._merge_hasher_factory()
-        for child in children:
-            check_file_or_directory(child)
-
-            if child.is_file():
-                header = _build_header(entry_name=child.name, entry_type="file")
-                hasher.update(header)
-                self._file_hasher.set_file(child)
-                digest = self._file_hasher.compute()
-                hasher.update(digest.digest_value)
-            else:
-                header = _build_header(entry_name=child.name, entry_type="dir")
-                hasher.update(header)
-                digest = self._dfs(child)
-                hasher.update(digest.digest_value)
-
-        return hasher.compute()
-
-
-def endpoints(step: int, end: int) -> Iterable[int]:
+def _endpoints(step: int, end: int) -> Iterable[int]:
     """Yields numbers from `step` to `end` inclusive, spaced by `step`.
 
     Last value is always equal to `end`, even when `end` is not a multiple of
@@ -155,11 +68,11 @@ def endpoints(step: int, end: int) -> Iterable[int]:
 
     Examples:
     ```python
-    >>> list(endpoints(2, 8))
+    >>> list(_endpoints(2, 8))
     [2, 4, 6, 8]
-    >>> list(endpoints(2, 9))
+    >>> list(_endpoints(2, 9))
     [2, 4, 6, 8, 9]
-    >>> list(endpoints(2, 2))
+    >>> list(_endpoints(2, 2))
     [2]
 
     Yields:
@@ -212,7 +125,7 @@ class ShardedDFSSerializer(serialization.Serializer):
 
         # TODO: github.com/sigstore/model-transparency/issues/196 - Add checks
         # to exclude symlinks if desired.
-        check_file_or_directory(model_path)
+        serialize_by_file.check_file_or_directory(model_path)
 
         if model_path.is_file():
             entries = [model_path]
@@ -282,13 +195,13 @@ class ShardedDFSSerializer(serialization.Serializer):
 
         tasks = []
         for path in paths:
-            check_file_or_directory(path)
+            serialize_by_file.check_file_or_directory(path)
             relative_path = path.relative_to(root_path)
 
             if path.is_file():
                 path_size = path.stat().st_size
                 start = 0
-                for end in endpoints(self._shard_size, path_size):
+                for end in _endpoints(self._shard_size, path_size):
                     tasks.append((relative_path, "file", start, end))
                     start = end
             else:
@@ -318,3 +231,118 @@ class ShardedDFSSerializer(serialization.Serializer):
         full_path = model_path.joinpath(task_path)
         hasher = self._file_hasher_factory(full_path, task_start, task_end)
         return hasher.compute().digest_value
+
+
+class ShardedFilesSerializer(serialization.Serializer):
+    """Model serializers that produces an itemized manifest, at shard level.
+
+    Traverses the model directory and creates digests for every file found,
+    sharding the file in equal shards and computing the digests in parallel.
+
+    Since the manifest lists each item individually, this will also enable
+    support for incremental updates (to be added later).
+    """
+
+    def __init__(
+        self,
+        sharded_hasher_factory: Callable[
+            [pathlib.Path, int, int], file.ShardedFileHasher
+        ],
+        max_workers: int | None = None,
+    ):
+        """Initializes an instance to serialize a model with this serializer.
+
+        Args:
+            sharded_hasher_factory: A callable to build the hash engine used to
+              hash every shard of the files in the model. Because each shard is
+              processed in parallel, every thread needs to call the factory to
+              start hashing. The arguments are the file, and the endpoints of
+              the shard.
+            max_workers: Maximum number of workers to use in parallel. Default
+              is to defer to the `concurrent.futures` library.
+        """
+        self._hasher_factory = sharded_hasher_factory
+        self._max_workers = max_workers
+
+        # Precompute some private values only once by using a mock file hasher.
+        # None of the arguments used to build the hasher are used.
+        hasher = sharded_hasher_factory(pathlib.Path(), 0, 1)
+        self._shard_size = hasher.shard_size
+
+    @override
+    def serialize(
+        self, model_path: pathlib.Path
+    ) -> manifest.ShardLevelManifest:
+        # TODO: github.com/sigstore/model-transparency/issues/196 - Add checks
+        # to exclude symlinks if desired.
+        serialize_by_file.check_file_or_directory(model_path)
+
+        shards = []
+        if model_path.is_file():
+            shards.extend(self._get_shards(model_path))
+        else:
+            # TODO: github.com/sigstore/model-transparency/issues/200 - When
+            # Python3.12 is the minimum supported version, this can be replaced
+            # with `pathlib.Path.walk` for a clearer interface, and some speed
+            # improvement.
+            for path in model_path.glob("**/*"):
+                serialize_by_file.check_file_or_directory(path)
+                if path.is_file():
+                    shards.extend(self._get_shards(path))
+
+        manifest_items = []
+        with concurrent.futures.ThreadPoolExecutor(
+            max_workers=self._max_workers
+        ) as tpe:
+            futures = [
+                tpe.submit(self._compute_hash, model_path, path, start, end)
+                for path, start, end in shards
+            ]
+            for future in concurrent.futures.as_completed(futures):
+                manifest_items.append(future.result())
+
+        return self._build_manifest(manifest_items)
+
+    def _get_shards(
+        self, path: pathlib.Path
+    ) -> list[tuple[pathlib.Path, int, int]]:
+        """Determines the shards of a given file path."""
+        shards = []
+        path_size = path.stat().st_size
+        if path_size > 0:
+            start = 0
+            for end in _endpoints(self._shard_size, path_size):
+                shards.append((path, start, end))
+                start = end
+        return shards
+
+    def _compute_hash(
+        self, model_path: pathlib.Path, path: pathlib.Path, start: int, end: int
+    ) -> manifest.ShardedFileManifestItem:
+        """Produces the manifest item of the file given by `path`.
+
+        Args:
+            model_path: The path to the model.
+            path: Path to the file in the model, that is currently transformed
+              to a manifest item.
+            start: The start offset of the shard (included).
+            end: The end offset of the shard (not included).
+
+        Returns:
+            The itemized manifest.
+        """
+        relative_path = path.relative_to(model_path)
+        digest = self._hasher_factory(path, start, end).compute()
+        return manifest.ShardedFileManifestItem(
+            path=relative_path, digest=digest, start=start, end=end
+        )
+
+    def _build_manifest(
+        self, items: Iterable[manifest.ShardedFileManifestItem]
+    ) -> manifest.ShardLevelManifest:
+        """Builds an itemized manifest from a given list of items.
+
+        Every subclass needs to implement this method to determine the format of
+        the manifest.
+        """
+        return manifest.ShardLevelManifest(items)
