+++ model_signing/serialization/serialize_by_file_shard_test.py
@@ -20,7 +20,7 @@ from model_signing.hashing import file
 from model_signing.hashing import memory
 from model_signing.manifest import manifest
 from model_signing.serialization import fixtures_constants
-from model_signing.serialization import itemized
+from model_signing.serialization import serialize_by_file_shard
 
 
 # Load fixtures from serialization/fixtures.py
@@ -52,22 +52,6 @@ def _extract_items_from_manifest(
     }
 
 
-def _extract_shard_items_from_manifest(
-    manifest: manifest.ShardLevelManifest,
-) -> dict[tuple[str, int, int], str]:
-    """Builds a dictionary representation of the items in a manifest.
-
-    Every item is mapped to its digest.
-
-    Used in multiple tests to check that we obtained the expected manifest.
-    """
-    return {
-        # convert to file path (relative to model) string and endpoints
-        (str(shard[0]), shard[1], shard[2]): digest.digest_hex
-        for shard, digest in manifest._item_to_digest.items()
-    }
-
-
 def _get_first_directory(path: pathlib.Path) -> pathlib.Path:
     """Returns the first directory that is a children of path.
 
@@ -84,24 +68,48 @@ def _get_first_file(path: pathlib.Path) -> pathlib.Path:
     return [f for f in path.iterdir() if f.is_file()][0]
 
 
-class TestFilesSerializer:
+class TestShardedDFSSerializer:
 
-    def _hasher_factory(self, path: pathlib.Path) -> file.FileHasher:
-        return file.SimpleFileHasher(path, memory.SHA256())
+    def _hasher_factory(
+        self, path: pathlib.Path, start: int, end: int
+    ) -> file.ShardedFileHasher:
+        return file.ShardedFileHasher(
+            path, memory.SHA256(), start=start, end=end
+        )
+
+    def _hasher_factory_small_shards(
+        self, path: pathlib.Path, start: int, end: int
+    ) -> file.ShardedFileHasher:
+        return file.ShardedFileHasher(
+            path, memory.SHA256(), start=start, end=end, shard_size=2
+        )
 
     def test_known_file(self, sample_model_file):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
-        expected = [
-            "3aab065c7181a173b5dd9e9d32a9f79923440b413be1e1ffcdba26a7365f719b"
-        ]
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
 
         manifest = serializer.serialize(sample_model_file)
-        digests = _extract_digests_from_manifest(manifest)
 
-        assert digests == expected
+        expected = (
+            "2ca48c47d5311a9b2f9305519cd5f927dcef09404fc32ef7886abe8f11450eff"
+        )
+        assert manifest.digest.digest_hex == expected
 
-    def test_file_manifest_unchanged_when_model_moved(self, sample_model_file):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
+    def test_file_hash_is_not_same_as_hash_of_content(self, sample_model_file):
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+
+        manifest = serializer.serialize(sample_model_file)
+
+        digest = memory.SHA256(fixtures_constants.KNOWN_MODEL_TEXT).compute()
+        assert manifest.digest.digest_hex != digest.digest_hex
+
+    def test_file_model_hash_is_same_if_model_is_moved(self, sample_model_file):
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
         manifest = serializer.serialize(sample_model_file)
 
         new_name = sample_model_file.with_name("new-file")
@@ -110,56 +118,53 @@ class TestFilesSerializer:
 
         assert manifest == new_manifest
 
-    def test_file_manifest_changes_if_content_changes(self, sample_model_file):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
+    def test_file_model_hash_changes_if_content_changes(
+        self, sample_model_file
+    ):
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
         manifest = serializer.serialize(sample_model_file)
-        digests = set(_extract_digests_from_manifest(manifest))
 
         sample_model_file.write_bytes(fixtures_constants.ANOTHER_MODEL_TEXT)
         new_manifest = serializer.serialize(sample_model_file)
-        new_digests = set(_extract_digests_from_manifest(new_manifest))
 
-        assert manifest != new_manifest
-        assert digests != new_digests
-        assert len(digests) == len(new_digests)
+        assert manifest.digest.algorithm == new_manifest.digest.algorithm
+        assert manifest.digest.digest_value != new_manifest.digest.digest_value
 
-    def test_directory_model_with_one_single_file(self, sample_model_file):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
-        manifest_file = serializer.serialize(sample_model_file)
-        digests_file = set(_extract_digests_from_manifest(manifest_file))
+    def test_directory_model_with_only_known_file(self, sample_model_file):
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
 
-        manifest = serializer.serialize(sample_model_file.parent)
-        digests = set(_extract_digests_from_manifest(manifest))
+        model = sample_model_file.parent
+        manifest = serializer.serialize(model)
 
-        assert manifest != manifest_file  # different paths
-        assert digests == digests_file
+        expected = (
+            "c030412c4c9e7f46396b591b1b6c4a4e40c15d9b9ca0b3512af8b20f3219c07f"
+        )
+        assert manifest.digest.digest_hex == expected
+        digest = memory.SHA256(fixtures_constants.KNOWN_MODEL_TEXT).compute()
+        assert manifest.digest.digest_hex != digest.digest_hex
 
     def test_known_folder(self, sample_model_folder):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
-        # Long hashes, want to update easily, so pylint: disable=line-too-long
-        expected_items = {
-            "f0": "997b37cc51f1ca1c7a270466607e26847429cd7264c30148c1b9352e224083fc",
-            "f1": "c88a04d48353133fb065ba2c8ab369abab21395b9526aa20373ad828915fa7ae",
-            "f2": "700e3ba5065d8dd47e41fd928ea086670d628f891ba363be0ca3c31d20d7d719",
-            "f3": "912bcf5ebdf44dc7b4085b07940e0a81d157fba24b276e73fd911121d4544c4a",
-            "d0/f00": "fdd8925354242a7fd1515e79534317b800015607a609cd306e0b4dcfe6c92249",
-            "d0/f01": "e16940b5e44ce981150bda37c4ba95881a749a521b4a297c5cdf97bdcfe965e6",
-            "d0/f02": "407822246ea8f9e26380842c3f4cd10d7b23e78f1fe7c74c293608682886a426",
-            "d1/f10": "6a3b08b5df77c4d418ceee1ac136a9ad49fc7c41358b5e82c1176daccb21ff3f",
-            "d1/f11": "a484b3d8ea5e99b75f9f123f9a42c882388693edc7d85d82ccba54834712cadf",
-            "d1/f12": "8f577930f5f40c2c2133cb299d36f9527fde98c1608569017cae6b5bcd01abb3",
-        }
-        # Re-enable lint, so pylint: enable=line-too-long
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
 
         manifest = serializer.serialize(sample_model_folder)
-        items = _extract_items_from_manifest(manifest)
 
-        assert items == expected_items
+        expected = (
+            "d22e0441cfa5ac2bc09715ddd88c802a7f97e29c93dc50f5498bab2954958ebb"
+        )
+        assert manifest.digest.digest_hex == expected
 
     def test_folder_model_hash_is_same_if_model_is_moved(
         self, sample_model_folder
     ):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
         manifest = serializer.serialize(sample_model_folder)
 
         new_name = sample_model_folder.with_name("new-root")
@@ -168,202 +173,181 @@ class TestFilesSerializer:
 
         assert manifest == new_manifest
 
-    def test_folder_model_empty_folder_not_included(self, sample_model_folder):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
-        manifest = serializer.serialize(sample_model_folder)
-
-        altered_dir = _get_first_directory(sample_model_folder)
-        new_empty_dir = altered_dir / "empty"
-        new_empty_dir.mkdir()
-        new_manifest = serializer.serialize(sample_model_folder)
-
-        assert manifest == new_manifest
-
-    def test_folder_model_empty_file_gets_included(self, sample_model_folder):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
-        manifest = serializer.serialize(sample_model_folder)
-
-        altered_dir = _get_first_directory(sample_model_folder)
-        new_empty_file = altered_dir / "empty"
-        new_empty_file.write_text("")
-        new_manifest = serializer.serialize(sample_model_folder)
-
-        assert manifest != new_manifest
-        assert (
-            len(new_manifest._item_to_digest)
-            == len(manifest._item_to_digest) + 1
-        )
-        for path, digest in manifest._item_to_digest.items():
-            assert new_manifest._item_to_digest[path] == digest
-
-    def _check_manifests_match_except_on_renamed_file(
-        self,
-        old_manifest: manifest.FileLevelManifest,
-        new_manifest: manifest.FileLevelManifest,
-        new_name: str,
-        old_name: pathlib.PurePath,
-    ):
-        """Checks that the manifests match, except on a renamed file.
-
-        For the renamed file, we still want to match the digest of the old name.
-        """
-        assert old_manifest != new_manifest
-        assert len(new_manifest._item_to_digest) == len(
-            old_manifest._item_to_digest
-        )
-        for path, digest in new_manifest._item_to_digest.items():
-            if path.name == new_name:
-                assert old_manifest._item_to_digest[old_name] == digest
-            else:
-                assert old_manifest._item_to_digest[path] == digest
-
-    def test_folder_model_rename_file_only_changes_path_part(
-        self, sample_model_folder
-    ):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
-        manifest = serializer.serialize(sample_model_folder)
-
-        altered_dir = _get_first_directory(sample_model_folder)
-        file_to_rename = _get_first_file(altered_dir)
-        old_name = file_to_rename.relative_to(sample_model_folder)
-        old_name = pathlib.PurePosixPath(old_name)  # canonicalize to Posix
-        new_name = file_to_rename.with_name("new-file")
-        file_to_rename.rename(new_name)
-        new_manifest = serializer.serialize(sample_model_folder)
-
-        self._check_manifests_match_except_on_renamed_file(
-            manifest, new_manifest, "new-file", old_name
-        )
-
-    def _check_manifests_match_except_on_renamed_dir(
-        self,
-        old_manifest: manifest.FileLevelManifest,
-        new_manifest: manifest.FileLevelManifest,
-        new_name: str,
-        old_name: str,
-    ):
-        """Checks that the manifests match, minus on paths under changed dir.
-
-        For paths under the changed directory, we want to match the digest of
-        the old path.
-        """
-        assert old_manifest != new_manifest
-        assert len(new_manifest._item_to_digest) == len(
-            old_manifest._item_to_digest
-        )
-        for path, digest in new_manifest._item_to_digest.items():
-            if new_name in path.parts:
-                parts = [
-                    old_name if part == new_name else part
-                    for part in path.parts
-                ]
-                old = pathlib.PurePosixPath(*parts)
-                assert old_manifest._item_to_digest[old] == digest
-            else:
-                assert old_manifest._item_to_digest[path] == digest
-
-    def test_folder_model_rename_dir_only_changes_path_part(
-        self, sample_model_folder
-    ):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
-        manifest = serializer.serialize(sample_model_folder)
-
-        dir_to_rename = _get_first_directory(sample_model_folder)
-        old_name = dir_to_rename.name
-        new_name = dir_to_rename.with_name("new-dir")
-        dir_to_rename.rename(new_name)
-        new_manifest = serializer.serialize(sample_model_folder)
-
-        self._check_manifests_match_except_on_renamed_dir(
-            manifest, new_manifest, "new-dir", old_name
-        )
-
-    def test_folder_model_replace_file_empty_folder(self, sample_model_folder):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
-        manifest = serializer.serialize(sample_model_folder)
-
-        altered_dir = _get_first_directory(sample_model_folder)
-        file_to_replace = _get_first_file(altered_dir)
-        file_to_replace.unlink()
-        file_to_replace.mkdir()
-        new_manifest = serializer.serialize(sample_model_folder)
-
-        assert manifest != new_manifest
-        assert (
-            len(new_manifest._item_to_digest)
-            == len(manifest._item_to_digest) - 1
-        )
-        for path, digest in new_manifest._item_to_digest.items():
-            assert manifest._item_to_digest[path] == digest
-
-    def _check_manifests_match_except_on_entry(
-        self,
-        old_manifest: manifest.FileLevelManifest,
-        new_manifest: manifest.FileLevelManifest,
-        expected_mismatch_path: pathlib.PurePath,
-    ):
-        """Checks that the manifests match, except for given path."""
-        assert old_manifest != new_manifest
-        assert len(new_manifest._item_to_digest) == len(
-            old_manifest._item_to_digest
-        )
-        for path, digest in new_manifest._item_to_digest.items():
-            if path == expected_mismatch_path:
-                assert old_manifest._item_to_digest[path] != digest
-            else:
-                assert old_manifest._item_to_digest[path] == digest
-
-    def test_folder_model_change_file(self, sample_model_folder):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
-        manifest = serializer.serialize(sample_model_folder)
-
-        altered_dir = _get_first_directory(sample_model_folder)
-        file_to_change = _get_first_file(altered_dir)
-        file_to_change.write_bytes(fixtures_constants.KNOWN_MODEL_TEXT)
-        changed_entry = file_to_change.relative_to(sample_model_folder)
-        changed_entry = pathlib.PurePosixPath(changed_entry)  # canonicalize
-        new_manifest = serializer.serialize(sample_model_folder)
-
-        self._check_manifests_match_except_on_entry(
-            manifest, new_manifest, changed_entry
-        )
-
-    def test_deep_folder(self, deep_model_folder):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
-        # Long hashes, want to update easily, so pylint: disable=line-too-long
-        expected_items = {
-            "d0/d1/d2/d3/d4/f0": "6efa14bb03544fcb76045c55f25b9315b6eb5be2d8a85f703193a76b7874c6ff",
-            "d0/d1/d2/d3/d4/f1": "a9bc149b70b9d325cd68d275d582cfdb98c0347d3ce54590aa6533368daed3d2",
-            "d0/d1/d2/d3/d4/f2": "5f597e6a92d1324d9adbed43d527926d11d0131487baf315e65ae1ef3b1ca3c0",
-            "d0/d1/d2/d3/d4/f3": "eaf677c35fec6b87889d9e4563d8bb65dcb9869ca0225697c9cc44cf49dca008",
-        }
-        # Re-enable lint, so pylint: enable=line-too-long
-
-        manifest = serializer.serialize(deep_model_folder)
-        items = _extract_items_from_manifest(manifest)
-
-        assert items == expected_items
-
     def test_empty_file(self, empty_model_file):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
-        expected = [
-            "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
-        ]
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
 
         manifest = serializer.serialize(empty_model_file)
-        digests = _extract_digests_from_manifest(manifest)
 
-        assert digests == expected
+        expected = (
+            "5f2d126b0d3540c17481fdf724e31cf03b4436a2ebabaa1d2e94fe09831be64d"
+        )
+        assert manifest.digest.digest_hex == expected
+
+    def test_directory_model_with_only_empty_file(self, empty_model_file):
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+        model = empty_model_file.parent
+
+        manifest = serializer.serialize(model)
+
+        expected = (
+            "74e81d0062f0a0674014c2f0e4b79985d5015f98a64089e7106a44d32e9ff11f"
+        )
+        assert manifest.digest.digest_hex == expected
 
     def test_empty_folder(self, empty_model_folder):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+
         manifest = serializer.serialize(empty_model_folder)
-        assert not manifest._item_to_digest
+
+        expected = (
+            "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855"
+        )
+        assert manifest.digest.digest_hex == expected
+
+    def test_folder_model_empty_entry(self, sample_model_folder):
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+
+        # Alter first directory within the model
+        dirs = [d for d in sample_model_folder.iterdir() if d.is_dir()]
+        altered_dir = dirs[0]
+
+        new_empty_dir = altered_dir / "empty"
+        new_empty_dir.mkdir()
+        manifest1 = serializer.serialize(sample_model_folder)
+
+        new_empty_dir.rmdir()
+
+        new_empty_file = altered_dir / "empty"
+        new_empty_file.write_text("")
+        manifest2 = serializer.serialize(sample_model_folder)
+
+        assert manifest1.digest != manifest2.digest
+
+    def test_folder_model_rename_file(self, sample_model_folder):
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+        manifest1 = serializer.serialize(sample_model_folder)
+
+        # Alter first directory within the model
+        dirs = [d for d in sample_model_folder.iterdir() if d.is_dir()]
+        altered_dir = dirs[0]
+
+        # Alter first file in the altered_dir
+        files = [f for f in altered_dir.iterdir() if f.is_file()]
+        file_to_rename = files[0]
+
+        new_name = file_to_rename.with_name("new-file")
+        file_to_rename.rename(new_name)
+
+        manifest2 = serializer.serialize(sample_model_folder)
+        assert manifest1.digest != manifest2.digest
+
+    def test_folder_model_rename_dir(self, sample_model_folder):
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+        manifest1 = serializer.serialize(sample_model_folder)
+
+        # Alter first directory within the model
+        dirs = [d for d in sample_model_folder.iterdir() if d.is_dir()]
+        dir_to_rename = dirs[0]
+
+        new_name = dir_to_rename.with_name("new-dir")
+        dir_to_rename.rename(new_name)
+
+        manifest2 = serializer.serialize(sample_model_folder)
+        assert manifest1.digest != manifest2.digest
+
+    def test_folder_model_replace_file_empty_folder(self, sample_model_folder):
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+        manifest1 = serializer.serialize(sample_model_folder)
+
+        # Alter first directory within the model
+        dirs = [d for d in sample_model_folder.iterdir() if d.is_dir()]
+        altered_dir = dirs[0]
+
+        # Replace first file in the altered_dir
+        files = [f for f in altered_dir.iterdir() if f.is_file()]
+        file_to_replace = files[0]
+        file_to_replace.unlink()
+        file_to_replace.mkdir()
+
+        manifest2 = serializer.serialize(sample_model_folder)
+        assert manifest1.digest != manifest2.digest
+
+    def test_folder_model_change_file(self, sample_model_folder):
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+        manifest1 = serializer.serialize(sample_model_folder)
+
+        # Alter first directory within the model
+        dirs = [d for d in sample_model_folder.iterdir() if d.is_dir()]
+        altered_dir = dirs[0]
+
+        # Alter first file in the altered_dir
+        files = [f for f in altered_dir.iterdir() if f.is_file()]
+        file_to_change = files[0]
+        file_to_change.write_bytes(fixtures_constants.KNOWN_MODEL_TEXT)
+
+        manifest2 = serializer.serialize(sample_model_folder)
+        assert manifest1.digest != manifest2.digest
+
+    def test_deep_folder(self, deep_model_folder):
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+
+        manifest = serializer.serialize(deep_model_folder)
+
+        expected = (
+            "52fa3c459aec58bc5f9702c73cb3c6b8fd19e9342aa3e4db851e1bde69ab1727"
+        )
+        assert manifest.digest.digest_hex == expected
+
+    def test_max_workers_does_not_change_digest(self, sample_model_folder):
+        serializer1 = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+        manifest1 = serializer1.serialize(sample_model_folder)
+
+        serializer2 = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256(), max_workers=2
+        )
+        manifest2 = serializer2.serialize(sample_model_folder)
+
+        assert manifest1 == manifest2
+
+    def test_shard_size_changes_digests(self, sample_model_folder):
+        serializer1 = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+        manifest1 = serializer1.serialize(sample_model_folder)
+
+        serializer2 = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory_small_shards, memory.SHA256()
+        )
+        manifest2 = serializer2.serialize(sample_model_folder)
+
+        assert manifest1.digest.digest_value != manifest2.digest.digest_value
 
     def test_special_file(self, sample_model_folder):
-        serializer = itemized.FilesSerializer(self._hasher_factory)
+        # Alter first directory within the model
+        dirs = [d for d in sample_model_folder.iterdir() if d.is_dir()]
+        altered_dir = dirs[0]
 
-        altered_dir = _get_first_directory(sample_model_folder)
+        # Create a pipe in the altered_dir
         pipe = altered_dir / "pipe"
 
         try:
@@ -372,31 +356,36 @@ class TestFilesSerializer:
             # On Windows, `os.mkfifo` does not exist (it should not).
             return  # trivially pass the test
 
+        serializer = serialize_by_file_shard.ShardedDFSSerializer(
+            self._hasher_factory, memory.SHA256()
+        )
+
         with pytest.raises(
             ValueError, match="Cannot use .* as file or directory"
         ):
             serializer.serialize(sample_model_folder)
 
+        # Also do the same for the pipe itself
         with pytest.raises(
             ValueError, match="Cannot use .* as file or directory"
         ):
             serializer.serialize(pipe)
 
-    def test_max_workers_does_not_change_digest(self, sample_model_folder):
-        serializer1 = itemized.FilesSerializer(self._hasher_factory)
-        serializer2 = itemized.FilesSerializer(
-            self._hasher_factory, max_workers=1
-        )
-        serializer3 = itemized.FilesSerializer(
-            self._hasher_factory, max_workers=3
-        )
 
-        manifest1 = serializer1.serialize(sample_model_folder)
-        manifest2 = serializer2.serialize(sample_model_folder)
-        manifest3 = serializer3.serialize(sample_model_folder)
+def _extract_shard_items_from_manifest(
+    manifest: manifest.ShardLevelManifest,
+) -> dict[tuple[str, int, int], str]:
+    """Builds a dictionary representation of the items in a manifest.
 
-        assert manifest1 == manifest2
-        assert manifest1 == manifest3
+    Every item is mapped to its digest.
+
+    Used in multiple tests to check that we obtained the expected manifest.
+    """
+    return {
+        # convert to file path (relative to model) string and endpoints
+        (str(shard[0]), shard[1], shard[2]): digest.digest_hex
+        for shard, digest in manifest._item_to_digest.items()
+    }
 
 
 class TestShardedFilesSerializer:
@@ -416,7 +405,9 @@ class TestShardedFilesSerializer:
         )
 
     def test_known_file(self, sample_model_file):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         expected = [
             "3aab065c7181a173b5dd9e9d32a9f79923440b413be1e1ffcdba26a7365f719b"
         ]
@@ -427,7 +418,9 @@ class TestShardedFilesSerializer:
         assert digests == expected
 
     def test_file_manifest_unchanged_when_model_moved(self, sample_model_file):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest = serializer.serialize(sample_model_file)
 
         new_name = sample_model_file.with_name("new-file")
@@ -437,7 +430,9 @@ class TestShardedFilesSerializer:
         assert manifest == new_manifest
 
     def test_file_manifest_changes_if_content_changes(self, sample_model_file):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest = serializer.serialize(sample_model_file)
         digests = set(_extract_digests_from_manifest(manifest))
 
@@ -450,7 +445,9 @@ class TestShardedFilesSerializer:
         assert digests != new_digests
 
     def test_directory_model_with_one_single_file(self, sample_model_file):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest_file = serializer.serialize(sample_model_file)
         digests_file = set(_extract_digests_from_manifest(manifest_file))
 
@@ -461,7 +458,9 @@ class TestShardedFilesSerializer:
         assert digests == digests_file
 
     def test_known_folder(self, sample_model_folder):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         # Long hashes, want to update easily, so pylint: disable=line-too-long
         expected_items = {
             (
@@ -525,7 +524,9 @@ class TestShardedFilesSerializer:
     def test_folder_model_hash_is_same_if_model_is_moved(
         self, sample_model_folder
     ):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest = serializer.serialize(sample_model_folder)
 
         new_name = sample_model_folder.with_name("new-root")
@@ -535,7 +536,9 @@ class TestShardedFilesSerializer:
         assert manifest == new_manifest
 
     def test_folder_model_empty_folder_not_included(self, sample_model_folder):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest = serializer.serialize(sample_model_folder)
 
         altered_dir = _get_first_directory(sample_model_folder)
@@ -546,7 +549,9 @@ class TestShardedFilesSerializer:
         assert manifest == new_manifest
 
     def test_folder_model_empty_file_not_included(self, sample_model_folder):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest = serializer.serialize(sample_model_folder)
 
         altered_dir = _get_first_directory(sample_model_folder)
@@ -582,7 +587,9 @@ class TestShardedFilesSerializer:
     def test_folder_model_rename_file_only_changes_path_part(
         self, sample_model_folder
     ):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest = serializer.serialize(sample_model_folder)
 
         altered_dir = _get_first_directory(sample_model_folder)
@@ -628,7 +635,9 @@ class TestShardedFilesSerializer:
     def test_folder_model_rename_dir_only_changes_path_part(
         self, sample_model_folder
     ):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest = serializer.serialize(sample_model_folder)
 
         dir_to_rename = _get_first_directory(sample_model_folder)
@@ -642,7 +651,9 @@ class TestShardedFilesSerializer:
         )
 
     def test_folder_model_replace_file_empty_folder(self, sample_model_folder):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest = serializer.serialize(sample_model_folder)
 
         altered_dir = _get_first_directory(sample_model_folder)
@@ -679,7 +690,9 @@ class TestShardedFilesSerializer:
                 assert old_manifest._item_to_digest[shard] == digest
 
     def test_folder_model_change_file(self, sample_model_folder):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest = serializer.serialize(sample_model_folder)
 
         altered_dir = _get_first_directory(sample_model_folder)
@@ -694,7 +707,9 @@ class TestShardedFilesSerializer:
         )
 
     def test_deep_folder(self, deep_model_folder):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         # Long hashes, want to update easily, so pylint: disable=line-too-long
         expected_items = {
             (
@@ -726,17 +741,23 @@ class TestShardedFilesSerializer:
         assert items == expected_items
 
     def test_empty_file(self, empty_model_file):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest = serializer.serialize(empty_model_file)
         assert not manifest._item_to_digest
 
     def test_empty_folder(self, empty_model_folder):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
         manifest = serializer.serialize(empty_model_folder)
         assert not manifest._item_to_digest
 
     def test_special_file(self, sample_model_folder):
-        serializer = itemized.ShardedFilesSerializer(self._hasher_factory)
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
 
         altered_dir = _get_first_directory(sample_model_folder)
         pipe = altered_dir / "pipe"
@@ -758,11 +779,13 @@ class TestShardedFilesSerializer:
             serializer.serialize(pipe)
 
     def test_max_workers_does_not_change_digest(self, sample_model_folder):
-        serializer1 = itemized.ShardedFilesSerializer(self._hasher_factory)
-        serializer2 = itemized.ShardedFilesSerializer(
+        serializer1 = serialize_by_file_shard.ShardedFilesSerializer(
+            self._hasher_factory
+        )
+        serializer2 = serialize_by_file_shard.ShardedFilesSerializer(
             self._hasher_factory, max_workers=1
         )
-        serializer3 = itemized.ShardedFilesSerializer(
+        serializer3 = serialize_by_file_shard.ShardedFilesSerializer(
             self._hasher_factory, max_workers=3
         )
 
@@ -774,7 +797,7 @@ class TestShardedFilesSerializer:
         assert manifest1 == manifest3
 
     def test_known_folder_small_shards(self, sample_model_folder):
-        serializer = itemized.ShardedFilesSerializer(
+        serializer = serialize_by_file_shard.ShardedFilesSerializer(
             self._hasher_factory_small_shards
         )
         # Long hashes, want to update easily, so pylint: disable=line-too-long
